Move SslConfigurationBuilder
