+++ modules/client/src/main/java/org/apache/ignite/client/SslConfiguration.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.client;
 
-import org.apache.ignite.internal.client.SslConfigurationBuilder;
+import org.apache.ignite.internal.client.SslConfigurationImpl;
 import org.jetbrains.annotations.Nullable;
 
 /** Client SSL configuration. */
@@ -41,7 +41,67 @@ public interface SslConfiguration {
     @Nullable String trustStorePassword();
 
     /** SSL configuration builder. */
-    static SslConfigurationBuilder builder() {
-        return new SslConfigurationBuilder();
+    static Builder builder() {
+        return new Builder();
+    }
+
+    @SuppressWarnings("PublicInnerClass")
+    class Builder {
+        private boolean enabled = false;
+
+        private @Nullable Iterable<String> ciphers;
+
+        private @Nullable String keyStorePath;
+
+        private @Nullable String keyStorePassword;
+
+        private @Nullable String trustStorePath;
+
+        private @Nullable String trustStorePassword;
+
+        /** Enabled/disabled setter. */
+        public Builder enabled(boolean enabled) {
+            this.enabled = enabled;
+            return this;
+        }
+
+        /** Ciphers setter. */
+        public Builder ciphers(@Nullable Iterable<String> ciphers) {
+            this.ciphers = ciphers;
+            return this;
+        }
+
+        /** Keystore path setter. */
+        public Builder keyStorePath(@Nullable String keyStorePath) {
+            this.keyStorePath = keyStorePath;
+            return this;
+        }
+
+        /** Keystore password setter. */
+        public Builder keyStorePassword(@Nullable String keyStorePassword) {
+            this.keyStorePassword = keyStorePassword;
+            return this;
+        }
+
+        /** Truststore path setter. */
+        public Builder trustStorePath(@Nullable String trustStorePath) {
+            this.trustStorePath = trustStorePath;
+            return this;
+        }
+
+        /** Truststore password setter. */
+        public Builder trustStorePassword(@Nullable String trustStorePassword) {
+            this.trustStorePassword = trustStorePassword;
+            return this;
+        }
+
+        /** Build SslConfiguration instance. */
+        public SslConfiguration build() {
+            return new SslConfigurationImpl(
+                    enabled, ciphers,
+                    keyStorePath, keyStorePassword,
+                    trustStorePath, trustStorePassword
+            );
+        }
     }
 }
